Create new struct CollectionIndentity; move id to CollectionIdentity; add guid to it
