+++ arangod/VocBase/Properties/CollectionInternalProperties.cpp
@@ -21,34 +21,3 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #include "CollectionInternalProperties.h"
-
-#include "Basics/NumberUtils.h"
-#include "Basics/Result.h"
-#include "VocBase/Properties/DatabaseConfiguration.h"
-#include "Inspection/Status.h"
-
-using namespace arangodb;
-
-inspection::Status
-CollectionInternalProperties::Transformers::IdIdentifier::toSerialized(
-    DataSourceId v, std::string& result) {
-  result = std::to_string(v.id());
-  return {};
-}
-
-inspection::Status
-CollectionInternalProperties::Transformers::IdIdentifier::fromSerialized(
-    std::string const& v, DataSourceId& result) {
-  char const* p = v.c_str();
-  result = DataSourceId{NumberUtils::atoi_zero<uint64_t>(p, p + v.length())};
-  return {};
-}
-
-[[nodiscard]] arangodb::Result
-CollectionInternalProperties::applyDefaultsAndValidateDatabaseConfiguration(
-    DatabaseConfiguration const& config) {
-  if (id.empty()) {
-    id = config.idGenerator();
-  }
-  return {TRI_ERROR_NO_ERROR};
-}
