+++ tests/Cluster/PlanCollectionToAgencyWriterTest.cpp
@@ -51,7 +51,7 @@ class PlanCollectionToAgencyWriterTest : public ::testing::Test {
         ->plan()
         ->collections()
         ->database(dbName())
-        ->collection(std::to_string(col.internal.id.id()))
+        ->collection(std::to_string(col.identity.id.id()))
         ->str();
   }
 
@@ -112,7 +112,7 @@ class PlanCollectionToAgencyWriterTest : public ::testing::Test {
       uint64_t nrShards, uint64_t replicationFactor) {
     CollectionDescriptor col{};
     col.mutableProps.name = "test";
-    col.internal.id = DataSourceId(123);
+    col.identity.id = DataSourceId(123);
     col.clusteringConstant.numberOfShards = nrShards;
     col.clusteringMutable.replicationFactor = replicationFactor;
     return col;
